[import perf] move definition selectors to prevent grpc import
