+++ python_modules/dagster-graphql/dagster_graphql/implementation/execution/backfill.py
@@ -3,11 +3,11 @@ from typing import TYPE_CHECKING, Any, List, Mapping, Union, cast
 import dagster._check as check
 import pendulum
 from dagster._core.definitions.external_asset_graph import ExternalAssetGraph
+from dagster._core.definitions.selector import RepositorySelector
 from dagster._core.errors import DagsterError
 from dagster._core.events import AssetKey
 from dagster._core.execution.backfill import BulkActionStatus, PartitionBackfill
 from dagster._core.execution.job_backfill import submit_backfill_runs
-from dagster._core.host_representation import RepositorySelector
 from dagster._core.utils import make_new_backfill_id
 from dagster._core.workspace.permissions import Permissions
