+++ python_modules/dagster-graphql/dagster_graphql/implementation/execution/launch_execution.py
@@ -1,8 +1,8 @@
 from typing import TYPE_CHECKING, cast
 
 import dagster._check as check
+from dagster._core.definitions.selector import PipelineSelector
 from dagster._core.execution.plan.resume_retry import ReexecutionStrategy
-from dagster._core.host_representation.selector import PipelineSelector
 from dagster._core.instance import DagsterInstance
 from dagster._core.storage.pipeline_run import DagsterRun, RunsFilter
 from dagster._core.workspace.permissions import Permissions
