+++ python_modules/dagster-graphql/dagster_graphql/implementation/external.py
@@ -5,8 +5,9 @@ from typing import TYPE_CHECKING, Mapping, Optional, Sequence, Union
 
 import dagster._check as check
 from dagster._config import validate_config_from_snap
+from dagster._core.definitions.selector import PipelineSelector, RepositorySelector
 from dagster._core.execution.plan.state import KnownExecutionState
-from dagster._core.host_representation import ExternalPipeline, PipelineSelector, RepositorySelector
+from dagster._core.host_representation import ExternalPipeline
 from dagster._core.host_representation.external import ExternalExecutionPlan
 from dagster._core.workspace.context import BaseWorkspaceRequestContext, WorkspaceRequestContext
 from dagster._utils.error import serializable_error_info_from_exc_info
