+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_instigators.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 import dagster._check as check
 from dagster._core.definitions.instigation_logger import get_instigation_log_records
 from dagster._core.definitions.run_request import InstigatorType
-from dagster._core.host_representation import InstigatorSelector
+from dagster._core.definitions.selector import InstigatorSelector
 from dagster._core.log_manager import DAGSTER_META_KEY
 from dagster._core.scheduler.instigation import InstigatorStatus
