+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_schedules.py
@@ -1,6 +1,10 @@
 import dagster._check as check
 from dagster._core.definitions.run_request import InstigatorType
-from dagster._core.host_representation import PipelineSelector, RepositorySelector, ScheduleSelector
+from dagster._core.definitions.selector import (
+    PipelineSelector,
+    RepositorySelector,
+    ScheduleSelector,
+)
 from dagster._core.workspace.permissions import Permissions
 from dagster._seven import get_current_datetime_in_utc, get_timestamp_from_utc_datetime
