+++ python_modules/dagster-graphql/dagster_graphql/implementation/utils.py
@@ -25,7 +25,7 @@ from typing import (
 
 import dagster._check as check
 from dagster._core.definitions.events import AssetKey
-from dagster._core.host_representation import GraphSelector, PipelineSelector
+from dagster._core.definitions.selector import GraphSelector, PipelineSelector
 from dagster._core.workspace.context import BaseWorkspaceRequestContext
 from dagster._utils.error import serializable_error_info_from_exc_info
 from typing_extensions import ParamSpec, TypeAlias
