+++ python_modules/dagster-test/dagster_test/test_project/__init__.py
@@ -7,6 +7,7 @@ from contextlib import contextmanager
 import dagster._check as check
 from dagster._core.code_pointer import FileCodePointer
 from dagster._core.definitions.reconstruct import ReconstructablePipeline, ReconstructableRepository
+from dagster._core.definitions.selector import InstigatorSelector
 from dagster._core.execution.api import create_execution_plan
 from dagster._core.execution.build_resources import build_resources
 from dagster._core.execution.context.output import build_output_context
@@ -15,7 +16,6 @@ from dagster._core.host_representation import (
     ExternalSchedule,
     GrpcServerRepositoryLocationOrigin,
     InProcessRepositoryLocationOrigin,
-    InstigatorSelector,
 )
 from dagster._core.host_representation.origin import (
     ExternalInstigatorOrigin,
