+++ python_modules/dagster/dagster/_core/execution/job_backfill.py
@@ -4,13 +4,13 @@ import time
 from typing import Iterable, Mapping, Optional, Sequence, Tuple, cast
 
 import dagster._check as check
+from dagster._core.definitions.selector import PipelineSelector
 from dagster._core.errors import DagsterBackfillFailedError
 from dagster._core.execution.plan.resume_retry import ReexecutionStrategy
 from dagster._core.execution.plan.state import KnownExecutionState
 from dagster._core.host_representation import (
     ExternalPartitionSet,
     ExternalPipeline,
-    PipelineSelector,
     RepositoryLocation,
 )
 from dagster._core.host_representation.external_data import (
