+++ python_modules/dagster/dagster/_core/scheduler/instigation.py
@@ -5,9 +5,11 @@ from typing import Any, Dict, List, Mapping, NamedTuple, Optional, Sequence, Typ
 import dagster._check as check
 
 # re-export
-from dagster._core.definitions.run_request import InstigatorType as InstigatorType
+from dagster._core.definitions.run_request import (
+    InstigatorType as InstigatorType,
+)
+from dagster._core.definitions.selector import InstigatorSelector, RepositorySelector
 from dagster._core.host_representation.origin import ExternalInstigatorOrigin
-from dagster._core.host_representation.selector import InstigatorSelector, RepositorySelector
 from dagster._serdes import create_snapshot_id
 from dagster._serdes.serdes import (
     DefaultNamedTupleSerializer,
