+++ python_modules/dagster/dagster/_core/test_utils.py
@@ -1,16 +1,13 @@
 import asyncio
 import os
 import re
-import sys
-import tempfile
 import time
 from collections import defaultdict
 from concurrent.futures import ThreadPoolExecutor
-from contextlib import ExitStack, contextmanager
-from typing import Any, Generator, Mapping, NamedTuple, Optional, Sequence, TypeVar
+from contextlib import contextmanager
+from typing import NamedTuple, Optional, Sequence, TypeVar
 
 import pendulum
-import yaml
 
 from dagster import (
     Permissive,
@@ -36,9 +33,14 @@ from dagster._legacy import ModeDefinition, pipeline, solid
 from dagster._serdes import ConfigurableClass
 from dagster._seven.compat.pendulum import create_pendulum_time, mock_pendulum_timezone
 from dagster._utils import Counter, get_terminate_signal, traced, traced_counter
-from dagster._utils.error import serializable_error_info_from_exc_info
 from dagster._utils.log import configure_loggers
-from dagster._utils.merger import merge_dicts
+
+# test utils from separate light weight file since are exported top level
+from .instance_for_test import (
+    cleanup_test_instance as cleanup_test_instance,
+    environ as environ,
+    instance_for_test as instance_for_test,
+)
 
 T_NamedTuple = TypeVar("T_NamedTuple", bound=NamedTuple)
 
@@ -97,111 +99,6 @@ def nesting_graph_pipeline(depth, num_children, *args, **kwargs):
         comp_solid.alias("outer")()
 
     return nested_pipeline
-
-
-@contextmanager
-def environ(env):
-    """Temporarily set environment variables inside the context manager and
-    fully restore previous environment afterwards.
-    """
-    previous_values = {key: os.getenv(key) for key in env}
-    for key, value in env.items():
-        if value is None:
-            if key in os.environ:
-                del os.environ[key]
-        else:
-            os.environ[key] = value
-    try:
-        yield
-    finally:
-        for key, value in previous_values.items():
-            if value is None:
-                if key in os.environ:
-                    del os.environ[key]
-            else:
-                os.environ[key] = value
-
-
-@contextmanager
-def instance_for_test(
-    overrides: Optional[Mapping[str, Any]] = None,
-    set_dagster_home: bool = True,
-    temp_dir: Optional[str] = None,
-) -> Generator[DagsterInstance, None, None]:
-    """Creates a persistent :py:class:`~dagster.DagsterInstance` available within a context manager.
-
-    When a context manager is opened, if no `temp_dir` parameter is set, a new
-    temporary directory will be created for the duration of the context
-    manager's opening. If the `set_dagster_home` parameter is set to True
-    (True by default), the `$DAGSTER_HOME` environment variable will be
-    overridden to be this directory (or the directory passed in by `temp_dir`)
-    for the duration of the context manager being open.
-
-    Args:
-        overrides (Optional[Mapping[str, Any]]):
-            Config to provide to instance (config format follows that typically found in an `instance.yaml` file).
-        set_dagster_home (Optional[bool]):
-            If set to True, the `$DAGSTER_HOME` environment variable will be
-            overridden to be the directory used by this instance for the
-            duration that the context manager is open. Upon the context
-            manager closing, the `$DAGSTER_HOME` variable will be re-set to the original value. (Defaults to True).
-        temp_dir (Optional[str]):
-            The directory to use for storing local artifacts produced by the
-            instance. If not set, a temporary directory will be created for
-            the duration of the context manager being open, and all artifacts
-            will be torn down afterward.
-    """
-    with ExitStack() as stack:
-        if not temp_dir:
-            temp_dir = stack.enter_context(tempfile.TemporaryDirectory())
-
-        # If using the default run launcher, wait for any grpc processes that created runs
-        # during test disposal to finish, since they might also be using this instance's tempdir
-        instance_overrides = merge_dicts(
-            {
-                "run_launcher": {
-                    "class": "DefaultRunLauncher",
-                    "module": "dagster._core.launcher.default_run_launcher",
-                    "config": {
-                        "wait_for_processes": True,
-                    },
-                },
-                "telemetry": {"enabled": False},
-            },
-            (overrides if overrides else {}),
-        )
-
-        if set_dagster_home:
-            stack.enter_context(
-                environ({"DAGSTER_HOME": temp_dir, "DAGSTER_DISABLE_TELEMETRY": "yes"})
-            )
-
-        with open(os.path.join(temp_dir, "dagster.yaml"), "w", encoding="utf8") as fd:
-            yaml.dump(instance_overrides, fd, default_flow_style=False)
-
-        with DagsterInstance.from_config(temp_dir) as instance:
-            try:
-                yield instance
-            except:
-                sys.stderr.write(
-                    "Test raised an exception, attempting to clean up instance:"
-                    + serializable_error_info_from_exc_info(sys.exc_info()).to_string()
-                    + "\n"
-                )
-                raise
-            finally:
-                cleanup_test_instance(instance)
-
-
-def cleanup_test_instance(instance):
-    check.inst_param(instance, "instance", DagsterInstance)
-    # To avoid filesystem contention when we close the temporary directory, wait for
-    # all runs to reach a terminal state, and close any subprocesses or threads
-    # that might be accessing the run history DB.
-
-    # Since launcher is lazy loaded, we don't need to do anyting if it's None
-    if instance._run_launcher:  # pylint: disable=protected-access
-        instance._run_launcher.join()  # pylint: disable=protected-access
 
 
 TEST_PIPELINE_NAME = "_test_pipeline_"
