+++ python_modules/dagster/dagster/_core/workspace/autodiscovery.py
@@ -12,7 +12,6 @@ from dagster._core.code_pointer import load_python_file, load_python_module
 from dagster._core.definitions import AssetGroup
 from dagster._core.definitions.definitions_class import Definitions
 from dagster._core.definitions.repository_definition import PendingRepositoryDefinition
-from dagster._legacy import PipelineDefinition
 
 LOAD_ALL_ASSETS = "<<LOAD_ALL_ASSETS>>"
 
@@ -54,6 +53,8 @@ def loadable_targets_from_python_package(
 
 
 def loadable_targets_from_loaded_module(module: ModuleType) -> Sequence[LoadableTarget]:
+    from dagster._legacy import PipelineDefinition
+
     loadable_defs = _loadable_targets_of_type(module, Definitions)
 
     if loadable_defs:
