+++ python_modules/dagster/dagster/_utils/alert.py
@@ -11,10 +11,10 @@ if TYPE_CHECKING:
     from dagster._core.definitions.graph_definition import GraphDefinition
     from dagster._core.definitions.pipeline_definition import PipelineDefinition
     from dagster._core.definitions.run_status_sensor_definition import RunFailureSensorContext
+    from dagster._core.definitions.selector import JobSelector, RepositorySelector
     from dagster._core.definitions.unresolved_asset_job_definition import (
         UnresolvedAssetJobDefinition,
     )
-    from dagster._core.host_representation.selector import JobSelector, RepositorySelector
 
 
 def _default_failure_email_body(context) -> str:
